course_diplom/requirements.txt
djangorestframework
djangorestframework-stubs
django-allauth
django-rest-auth
pyyaml
drf-nested-routers